in-progress changes - print changes
making some changes to how the print statements are done so the console is in a more readable state.

Note: For better practice these should be moved into their own class so the methods may be called in any other classes -> in this case it's not needed but would be a good idea for real projects.